fix(transient): Initialization not removing suffixes that have been deleted from the initial state
Since the state of the prefix is saved through `transient-set' when the prefix
initialized after a deletion of an argument through `blue-transient--del', the
state still contains the saved values of the suffixes for the selected
command. What the prefix should initialized to is to the state defined by
`blue-transient--command-chain' so this commit makes sure that those saved
suffixes are filtered out if they are no longer present in the chain.