perf: Improve startup of commands by running external blue processes in build directory
When using an embedded BLUE through `BLUE_EMBEDDED', BLUE adds that directory to
the load path and the current working directory to the compiled load path;
unless, `BLUE_EMBEDDED' has the form of `load-path:load-compiled-path'.

The UI can take advantage of the fact that the CWD will be added to the compiled
path and try to execute any external BLUE process under a known build directory,
this would make use of the embedded compiled modules if available, greatly
improving the startup time.