fix: Reevaluation of 'blue-transient.el' leads to inconsistent state of `blue-transient--menu'
The original definition was a stub to silence the byte-compilation warning for
the undeclared function `blue-transient--menu'. That function is defined during
the evaluation of `blue-transient'. The problem is that if 'blue-transient.el'
is reevaluated, the function gets redefined to it's original stub but the
`blue-transient--menu-expresion' variable does not get reset to nil. This leads
to `blue-transient' not redefining the stub so it does nothing.